+++ yue-library-base/src/main/java/ai/yue/library/base/util/AESUtils.java
@@ -16,10 +16,12 @@ import lombok.extern.slf4j.Slf4j;
 /**
  * AES 加密工具类
  * 
+ * @deprecated 请使用 yue-library-base-crypto 模块 SecureSingleton
  * @author	孙金川
  * @since	2018年3月28日
  */
 @Slf4j
+@Deprecated
 public class AESUtils {
 
 	private static final String KEY_ALGORITHM = "AES";