+++ yue-library-base/src/main/java/ai/yue/library/base/util/URIUtils.java
@@ -68,10 +68,13 @@ public class URIUtils {
 	/**
 	 * 1. 将URI转义内容进行解码<br>
 	 * 2. 将RSA 1024分段加密内容，进行分段解密
+	 * 
+	 * @deprecated 请使用 yue-library-base-crypto 模块 SecureSingleton
 	 * @param text				URI转义后的文本
 	 * @param rsa_private_key	RSA1024私钥
 	 * @return 解密后的JSON
 	 */
+	@Deprecated
 	public static JSONObject rsaUriDecodingAndDecrypt(String text, String rsa_private_key) {
 		String content = decode(text);
 		String jsonString = RSAUtils.decrypt(content, rsa_private_key);